Create README.md
This repository contains all the input data in Abaqus for my Master's thesis: Fracture behavior of concrete under uniaxial tension.

These pieces of code employ the finite element method implemented with Abaqus to simulate the responses of concrete in direct tension test with different influence factors. Several models are constructed with different end-constraints, material properties and dimensions.

It must be admitted that the simulations are imperfect due to my deficiency in knowledge and experience, so any comments or revising suggestions are desirably welcome.